Conflicts:
	entity-management-web/src/main/java/eu/europeana/entitymanagement/web/EntityAdminController.java

+++ entity-management-web/src/main/java/eu/europeana/entitymanagement/batch/writer/EntityRecordDatabaseRemovalWriter.java
@@ -1,16 +1,22 @@
 package eu.europeana.entitymanagement.batch.writer;
 
 import eu.europeana.entitymanagement.batch.utils.BatchUtils;
-import eu.europeana.entitymanagement.definitions.model.EntityRecord;
+import eu.europeana.entitymanagement.definitions.batch.model.BatchEntityRecord;
+import eu.europeana.entitymanagement.definitions.batch.model.ScheduledRemovalType;
+import eu.europeana.entitymanagement.definitions.batch.model.ScheduledTaskType;
 import eu.europeana.entitymanagement.web.service.EntityRecordService;
 import java.util.List;
+import java.util.Set;
 import org.springframework.batch.item.ItemWriter;
 import org.springframework.lang.NonNull;
 import org.springframework.stereotype.Component;
 
 /** ItemWriter for deleting EntityRecords from Mongo */
 @Component
-public class EntityRecordDatabaseRemovalWriter implements ItemWriter<EntityRecord> {
+public class EntityRecordDatabaseRemovalWriter implements ItemWriter<BatchEntityRecord> {
+
+  private static final Set<ScheduledTaskType> supportedScheduledTasks =
+      Set.of(ScheduledRemovalType.PERMANENT_DELETION);
 
   private final EntityRecordService entityRecordService;
 
@@ -19,8 +25,11 @@ public class EntityRecordDatabaseRemovalWriter implements ItemWriter<EntityRecor
   }
 
   @Override
-  public void write(@NonNull List<? extends EntityRecord> entityRecords) throws Exception {
-    List<String> ids = List.of(BatchUtils.getEntityIds(entityRecords));
-    entityRecordService.deleteBulk(ids, false);
+  public void write(@NonNull List<? extends BatchEntityRecord> entityRecords) throws Exception {
+    List<String> ids = BatchUtils.filterRecordsForWriters(supportedScheduledTasks, entityRecords);
+
+    if (!ids.isEmpty()) {
+      entityRecordService.deleteBulk(ids, false);
+    }
   }
 }
